refactor(refs T30969): move DpSelect to components root
