+++ src/components/DpSelect/DpSelect.vue
@@ -35,7 +35,7 @@
 </template>
 
 <script>
-import DpLabel from '../../DpLabel/DpLabel'
+import DpLabel from '../DpLabel/DpLabel'
 import { prefixClassMixin } from '@demos-europe/demosplan-utils'
 
 export default {
